requirements.txt
flask
vanna[snowflake]
db-dtypes
python-dotenv
openai
chromadb
vanna[postgres]
WSGIServer
gevent
flask-cors